requirements.txt
# Core
Django==5.2.1
djangorestframework==3.15.2
# Token auth comes with DRF; no extra package needed

# Images & PDFs
Pillow==10.4.0
reportlab==4.2.2

# Dev helpers (optional, comment out if not needed)
# django-cors-headers==4.4.0
